% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaeData-pkg.R
\name{demo_dir_file}
\alias{demo_dir_file}
\title{Split getter output}
\usage{
demo_dir_file(ehub_dir, dir = TRUE)
}
\arguments{
\item{ehub_dir}{character string that is retrieved by the getters}

\item{dir}{binary if the output should contain file path and file name or only file name}
}
\value{
list containing (file path) and file name
}
\description{
Internal function used by \code{ehoub_out()} that splits the output from the getter into a file path and the corresponding file name.
This is necessary as the read-in function \code{read_allele_counts()} from the \code{SingleCellAlleleExperiment} package
expects a directory path as well as the names of each expected file.
}
