% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaeData-pkg.R
\name{ehub_out}
\alias{ehub_out}
\title{Build output list per dataset}
\usage{
ehub_out(bc_dir, feature_dir, mtx_dir)
}
\arguments{
\item{bc_dir}{character string containing full path to barcode file}

\item{feature_dir}{character string containing full path to feature file}

\item{mtx_dir}{character string containing full path to matrix file}
}
\value{
list with four elements containing file path to the directory containing all files and each file name for barcodes, features and matrix
}
\description{
Internal function used by \code{scae_data()} to build a list containing the file path
}
