% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaeData.R
\name{get_barcodes_20k}
\alias{get_barcodes_20k}
\title{Get barcode identifiers for \verb{pbmc-20k} dataset}
\usage{
get_barcodes_20k()
}
\value{
character string containing the file path of the barcode file
}
\description{
Internal getter function that retrieves the \code{pbmc_20k_barcodes.txt} from \code{ExperimentHub (eh[["EH9459"]])}. This is one out of 3
files that are necessary to load for using the data in a \emph{\href{https://github.com/AGImkeller/SingleCellAlleleExperiment}{SingleCellAlleleExperiment}} data structure.
File is generated with the \emph{\href{https://github.com/AGImkeller/scIGD}{scIGD}} workflow.
}
