% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaeData-pkg.R
\name{get_counts_5k}
\alias{get_counts_5k}
\title{Get quantification matrix for \verb{pbmc-5k} dataset}
\usage{
get_counts_5k()
}
\value{
character string containing the file path of the matrix file
}
\description{
Internal getter function that retrieves the \code{pbmc_5k_count_mtx.mtx} from \code{ExperimentHub (eh[["EH9455"]])}. This is one out of 3
files that are necessary to load for using the data in a \emph{\href{https://github.com/AGImkeller/SingleCellAlleleExperiment}{SingleCellAlleleExperiment}} data structure.
File is generated with the \emph{\href{https://github.com/AGImkeller/scIGD}{scIGD}} workflow.
}
