% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaeData-pkg.R
\docType{package}
\name{scaeData-pkg}
\alias{scaeData}
\alias{scaeData-package}
\alias{scaeData-pkg}
\title{scaeData}
\description{
scaeData: Data Package for SingleCellAlleleExperiment
}
\details{
This package contains some exemplary datasets used by the Bioconductor package
\code{SingleCellAlleleExperiment}.
The raw FASTQ files were sourced from publicly accessible datasets provided by
10x Genomics.
Subsequently, our \code{scIGD} snakemake workflow was employed to process these
FASTQ files in order to obtain allele-level quantifications.
The resulting output from \code{scIGD} constitutes to the contents of this data
package.
}
\seealso{
https://github.com/AGImkeller/scIGD/ for the definition of the
quantification workflow.
}
\author{
\strong{Maintainer}: Ahmad Al Ajami \email{alajami@med.uni-frankfurt.de} (\href{https://orcid.org/0009-0006-5615-7447}{ORCID})

Authors:
\itemize{
  \item Jonas Schuck \email{schuck@med.uni-frankfurt.de} (\href{https://orcid.org/0009-0003-5705-4579}{ORCID})
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
  \item Katharina Imkeller \email{imkeller@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-5177-0852}{ORCID})
}

}
\keyword{internal}
